+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v2.2.0 for all versions of KSP from 1.3.1 to 1.6.x
 
- - 2019-xx-xx
+ - 2019-01-02
 
 ### Changes since the last release
 
@@ -9,6 +9,7 @@
 * Processes default to a switched on state when added in the VAB/SPH (PiezPiedPy) 
 * Made the PartUpgrade for module slots require ProfileDefault (theJesuit)
 * Took away some of the Partupgrade as I upgraded my MM fu. (theJesuit)
+* Fixed compatability with Module Manager 4.x (steamp0rt, with lots of help from blowfish)
 
 ------------------------------------------------------------------------------------------------------